requirements.txt
streamlit
boto3
pandas
plotly
yfinance
kafka-python
s3fs
ta
streamlit-authenticator